[Project 4-2] Add full disk detection at inode_create
