+++ filesys/inode.c
@@ -87,6 +87,9 @@ inode_create (disk_sector_t sector, off_t length, bool isdir) {
 
 	ASSERT (length >= 0);
 
+	if(sector >= filesys_disk->capacity)
+		return false;
+
 	/* If this assertion fails, the inode structure is not exactly
 	 * one sector in size, and you should fix that. */
 	ASSERT (sizeof *disk_inode == DISK_SECTOR_SIZE);
